requirements.txt
confluent-kafka>=2.4.0
psycopg2-binary>=2.9.9
pytest>=8.0.0
requests>=2.31.0
